requirements.txt
nnsight==0.4.0
jupyter==1.0.0
matplotlib==3.9.2
numpy==1.26.2
pandas==2.1.3
pytorch==2.4.1
openai==1.6.1
einops==0.8.0
rich==13.9.4
scipy==1.11.4
seaborn==0.12.2
tqdm==4.66.4